git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@18831 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/magic.c
@@ -249,9 +249,9 @@ static int CompareMagickInfoSize(const void *a,const void *b)
     *ma,
     *mb;
 
-  ma=(MagicInfo *)a;
-  mb=(MagicInfo *)b;
-  return((mb->offset+mb->length) - (ma->offset+ma->length));
+  ma=(MagicInfo *) a;
+  mb=(MagicInfo *) b;
+  return((int) ((mb->offset+mb->length)-(ma->offset+ma->length)));
 }
 
 static LinkedListInfo *AcquireMagicCache(const char *filename,
